Work on integrating character with home
