+++ Icarus/src/main/java/com/revature/data/CharacterDAO.java
@@ -4,7 +4,7 @@ import java.util.Set;
 
 import com.revature.beans.Character;
 public interface CharacterDAO {
-	public Character addUser(Character character);
+	public Character addCharacter(Character character);
 	public Character getCharacterById(int characterId);
 	public Set<Character> getAllCharacters();
 	public Set<Character> getAllCharactersByUser(int userId);
